+++ api/incubator/src/main/java/io/opentelemetry/api/incubator/config/DeclarativeConfigProperties.java
@@ -3,20 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure.spi.internal;
+package io.opentelemetry.api.incubator.config;
 
 import static io.opentelemetry.api.internal.ConfigUtil.defaultIfNull;
 
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigurationException;
 import java.util.List;
 import java.util.Set;
 import javax.annotation.Nullable;
 
 /**
- * An interface for accessing structured configuration data.
+ * An interface for accessing declarative configuration data.
  *
- * <p>An instance of {@link StructuredConfigProperties} is equivalent to a <a
+ * <p>An instance of {@link DeclarativeConfigProperties} is equivalent to a <a
  * href="https://yaml.org/spec/1.2.2/#3211-nodes">YAML mapping node</a>. It has accessors for
  * reading scalar properties, {@link #getStructured(String)} for reading children which are
  * themselves mappings, and {@link #getStructuredList(String)} for reading children which are
@@ -25,24 +23,24 @@ import javax.annotation.Nullable;
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
-public interface StructuredConfigProperties {
+public interface DeclarativeConfigProperties {
 
   /**
-   * Return an empty {@link StructuredConfigProperties} instance.
+   * Return an empty {@link DeclarativeConfigProperties} instance.
    *
    * <p>Useful for walking the tree without checking for null. For example, to access a string key
    * nested at .foo.bar.baz, call: {@code config.getStructured("foo", empty()).getStructured("bar",
    * empty()).getString("baz")}.
    */
-  static StructuredConfigProperties empty() {
-    return EmptyStructuredConfigProperties.getInstance();
+  static DeclarativeConfigProperties empty() {
+    return EmptyDeclarativeConfigProperties.getInstance();
   }
 
   /**
    * Returns a {@link String} configuration property.
    *
    * @return null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar string
+   * @throws DeclarativeConfigException if the property is not a valid scalar string
    */
   @Nullable
   String getString(String name);
@@ -52,7 +50,7 @@ public interface StructuredConfigProperties {
    *
    * @return a {@link String} configuration property or {@code defaultValue} if a property with
    *     {@code name} has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar string
+   * @throws DeclarativeConfigException if the property is not a valid scalar string
    */
   default String getString(String name, String defaultValue) {
     return defaultIfNull(getString(name), defaultValue);
@@ -63,7 +61,7 @@ public interface StructuredConfigProperties {
    * {@link Boolean#parseBoolean(String)} for handling the values.
    *
    * @return null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar boolean
+   * @throws DeclarativeConfigException if the property is not a valid scalar boolean
    */
   @Nullable
   Boolean getBoolean(String name);
@@ -73,7 +71,7 @@ public interface StructuredConfigProperties {
    *
    * @return a {@link Boolean} configuration property or {@code defaultValue} if a property with
    *     {@code name} has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar boolean
+   * @throws DeclarativeConfigException if the property is not a valid scalar boolean
    */
   default boolean getBoolean(String name, boolean defaultValue) {
     return defaultIfNull(getBoolean(name), defaultValue);
@@ -86,7 +84,7 @@ public interface StructuredConfigProperties {
    * {@link Long#intValue()} which may result in loss of precision.
    *
    * @return null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar integer
+   * @throws DeclarativeConfigException if the property is not a valid scalar integer
    */
   @Nullable
   Integer getInt(String name);
@@ -99,7 +97,7 @@ public interface StructuredConfigProperties {
    *
    * @return a {@link Integer} configuration property or {@code defaultValue} if a property with
    *     {@code name} has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar integer
+   * @throws DeclarativeConfigException if the property is not a valid scalar integer
    */
   default int getInt(String name, int defaultValue) {
     return defaultIfNull(getInt(name), defaultValue);
@@ -109,7 +107,7 @@ public interface StructuredConfigProperties {
    * Returns a {@link Long} configuration property.
    *
    * @return null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar long
+   * @throws DeclarativeConfigException if the property is not a valid scalar long
    */
   @Nullable
   Long getLong(String name);
@@ -119,7 +117,7 @@ public interface StructuredConfigProperties {
    *
    * @return a {@link Long} configuration property or {@code defaultValue} if a property with {@code
    *     name} has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar long
+   * @throws DeclarativeConfigException if the property is not a valid scalar long
    */
   default long getLong(String name, long defaultValue) {
     return defaultIfNull(getLong(name), defaultValue);
@@ -129,7 +127,7 @@ public interface StructuredConfigProperties {
    * Returns a {@link Double} configuration property.
    *
    * @return null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar double
+   * @throws DeclarativeConfigException if the property is not a valid scalar double
    */
   @Nullable
   Double getDouble(String name);
@@ -139,7 +137,7 @@ public interface StructuredConfigProperties {
    *
    * @return a {@link Double} configuration property or {@code defaultValue} if a property with
    *     {@code name} has not been configured
-   * @throws ConfigurationException if the property is not a valid scalar double
+   * @throws DeclarativeConfigException if the property is not a valid scalar double
    */
   default double getDouble(String name, double defaultValue) {
     return defaultIfNull(getDouble(name), defaultValue);
@@ -153,8 +151,8 @@ public interface StructuredConfigProperties {
    * @param scalarType the scalar type, one of {@link String}, {@link Boolean}, {@link Long} or
    *     {@link Double}
    * @return a {@link List} configuration property, or null if the property has not been configured
-   * @throws ConfigurationException if the property is not a valid sequence of scalars, or if {@code
-   *     scalarType} is not supported
+   * @throws DeclarativeConfigException if the property is not a valid sequence of scalars, or if
+   *     {@code scalarType} is not supported
    */
   @Nullable
   <T> List<T> getScalarList(String name, Class<T> scalarType);
@@ -163,56 +161,58 @@ public interface StructuredConfigProperties {
    * Returns a {@link List} configuration property. Entries which are not strings are converted to
    * their string representation.
    *
-   * @see ConfigProperties#getList(String name)
+   * @param name the property name
+   * @param scalarType the scalar type, one of {@link String}, {@link Boolean}, {@link Long} or
+   *     {@link Double}
    * @return a {@link List} configuration property or {@code defaultValue} if a property with {@code
    *     name} has not been configured
-   * @throws ConfigurationException if the property is not a valid sequence of scalars
+   * @throws DeclarativeConfigException if the property is not a valid sequence of scalars
    */
   default <T> List<T> getScalarList(String name, Class<T> scalarType, List<T> defaultValue) {
     return defaultIfNull(getScalarList(name, scalarType), defaultValue);
   }
 
   /**
-   * Returns a {@link StructuredConfigProperties} configuration property.
+   * Returns a {@link DeclarativeConfigProperties} configuration property.
    *
    * @return a map-valued configuration property, or {@code null} if {@code name} has not been
    *     configured
-   * @throws ConfigurationException if the property is not a mapping
+   * @throws DeclarativeConfigException if the property is not a mapping
    */
   @Nullable
-  StructuredConfigProperties getStructured(String name);
+  DeclarativeConfigProperties getStructured(String name);
 
   /**
-   * Returns a {@link StructuredConfigProperties} configuration property.
+   * Returns a list of {@link DeclarativeConfigProperties} configuration property.
    *
    * @return a map-valued configuration property, or {@code defaultValue} if {@code name} has not
    *     been configured
-   * @throws ConfigurationException if the property is not a mapping
+   * @throws DeclarativeConfigException if the property is not a mapping
    */
-  default StructuredConfigProperties getStructured(
-      String name, StructuredConfigProperties defaultValue) {
+  default DeclarativeConfigProperties getStructured(
+      String name, DeclarativeConfigProperties defaultValue) {
     return defaultIfNull(getStructured(name), defaultValue);
   }
 
   /**
-   * Returns a list of {@link StructuredConfigProperties} configuration property.
+   * Returns a list of {@link DeclarativeConfigProperties} configuration property.
    *
    * @return a list of map-valued configuration property, or {@code null} if {@code name} has not
    *     been configured
-   * @throws ConfigurationException if the property is not a sequence of mappings
+   * @throws DeclarativeConfigException if the property is not a sequence of mappings
    */
   @Nullable
-  List<StructuredConfigProperties> getStructuredList(String name);
+  List<DeclarativeConfigProperties> getStructuredList(String name);
 
   /**
-   * Returns a list of {@link StructuredConfigProperties} configuration property.
+   * Returns a list of {@link DeclarativeConfigProperties} configuration property.
    *
    * @return a list of map-valued configuration property, or {@code defaultValue} if {@code name}
    *     has not been configured
-   * @throws ConfigurationException if the property is not a sequence of mappings
+   * @throws DeclarativeConfigException if the property is not a sequence of mappings
    */
-  default List<StructuredConfigProperties> getStructuredList(
-      String name, List<StructuredConfigProperties> defaultValue) {
+  default List<DeclarativeConfigProperties> getStructuredList(
+      String name, List<DeclarativeConfigProperties> defaultValue) {
     return defaultIfNull(getStructuredList(name), defaultValue);
   }