+++ exporters/otlp/testing-internal/src/main/java/io/opentelemetry/exporter/otlp/testing/internal/GrpcLogRecordExporterBuilderWrapper.java
@@ -12,6 +12,7 @@ import io.opentelemetry.sdk.common.export.RetryPolicy;
 import io.opentelemetry.sdk.logs.data.LogRecordData;
 import java.time.Duration;
 import java.util.Map;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
@@ -109,6 +110,20 @@ final class GrpcLogRecordExporterBuilderWrapper implements TelemetryExporterBuil
   @SuppressWarnings("deprecation") // testing deprecated functionality
   public TelemetryExporterBuilder<LogRecordData> setChannel(Object channel) {
     builder.setChannel((ManagedChannel) channel);
+    return this;
+  }
+
+  @Override
+  public TelemetryExporterBuilder<LogRecordData> setServiceClassLoader(
+      ClassLoader serviceClassLoader) {
+    builder.setServiceClassLoader(serviceClassLoader);
+    return this;
+  }
+
+  @Override
+  public TelemetryExporterBuilder<LogRecordData> setExecutorService(
+      ExecutorService executorService) {
+    builder.setExecutorService(executorService);
     return this;
   }