+++ sdk-extensions/incubator/src/test/java/io/opentelemetry/sdk/extension/incubator/fileconfig/DeclarativeConfigurationParseTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.sdk.extension.incubator.fileconfig;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigurationException;
+import io.opentelemetry.api.incubator.config.DeclarativeConfigException;
 import io.opentelemetry.sdk.extension.incubator.fileconfig.internal.model.AggregationModel;
 import io.opentelemetry.sdk.extension.incubator.fileconfig.internal.model.AlwaysOffModel;
 import io.opentelemetry.sdk.extension.incubator.fileconfig.internal.model.AlwaysOnModel;
@@ -76,15 +76,15 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class FileConfigurationParseTest {
+class DeclarativeConfigurationParseTest {
 
   @Test
   void parse_BadInputStream() {
     assertThatThrownBy(
             () ->
-                FileConfiguration.parseAndCreate(
+                DeclarativeConfiguration.parseAndCreate(
                     new ByteArrayInputStream("foo".getBytes(StandardCharsets.UTF_8))))
-        .isInstanceOf(ConfigurationException.class)
+        .isInstanceOf(DeclarativeConfigException.class)
         .hasMessage("Unable to parse configuration input stream");
   }
 
@@ -446,7 +446,7 @@ class FileConfigurationParseTest {
 
     try (FileInputStream configExampleFile =
         new FileInputStream(System.getenv("CONFIG_EXAMPLE_DIR") + "/kitchen-sink.yaml")) {
-      OpenTelemetryConfigurationModel config = FileConfiguration.parse(configExampleFile);
+      OpenTelemetryConfigurationModel config = DeclarativeConfiguration.parse(configExampleFile);
 
       // General config
       assertThat(config.getFileFormat()).isEqualTo("0.3");
@@ -499,7 +499,7 @@ class FileConfigurationParseTest {
             + "        aggregation:\n"
             + "          drop: {}\n";
     OpenTelemetryConfigurationModel objectPlaceholderModel =
-        FileConfiguration.parse(
+        DeclarativeConfiguration.parse(
             new ByteArrayInputStream(objectPlaceholderString.getBytes(StandardCharsets.UTF_8)));
 
     String noOjbectPlaceholderString =
@@ -517,7 +517,7 @@ class FileConfigurationParseTest {
             + "        aggregation:\n"
             + "          drop:\n";
     OpenTelemetryConfigurationModel noObjectPlaceholderModel =
-        FileConfiguration.parse(
+        DeclarativeConfiguration.parse(
             new ByteArrayInputStream(noOjbectPlaceholderString.getBytes(StandardCharsets.UTF_8)));
 
     SpanExporterModel exporter =
@@ -551,7 +551,8 @@ class FileConfigurationParseTest {
             + "      ratio:\n"; // Double
 
     OpenTelemetryConfigurationModel model =
-        FileConfiguration.parse(new ByteArrayInputStream(yaml.getBytes(StandardCharsets.UTF_8)));
+        DeclarativeConfiguration.parse(
+            new ByteArrayInputStream(yaml.getBytes(StandardCharsets.UTF_8)));
 
     assertThat(model.getFileFormat()).isNull();
     assertThat(model.getDisabled()).isNull();
@@ -573,7 +574,7 @@ class FileConfigurationParseTest {
   @MethodSource("coreSchemaValuesArgs")
   void coreSchemaValues(String rawYaml, Object expectedYamlResult) {
     Object yaml =
-        FileConfiguration.loadYaml(
+        DeclarativeConfiguration.loadYaml(
             new ByteArrayInputStream(rawYaml.getBytes(StandardCharsets.UTF_8)),
             Collections.emptyMap());
     assertThat(yaml).isEqualTo(expectedYamlResult);
@@ -601,7 +602,7 @@ class FileConfigurationParseTest {
     environmentVariables.put("HEX", "0xdeadbeef");
 
     Object yaml =
-        FileConfiguration.loadYaml(
+        DeclarativeConfiguration.loadYaml(
             new ByteArrayInputStream(rawYaml.getBytes(StandardCharsets.UTF_8)),
             environmentVariables);
     assertThat(yaml).isEqualTo(expectedYamlResult);
@@ -689,7 +690,7 @@ class FileConfigurationParseTest {
     Map<String, String> envVars = new HashMap<>();
     envVars.put("OTEL_EXPORTER_OTLP_ENDPOINT", "http://collector:4317");
     OpenTelemetryConfigurationModel model =
-        FileConfiguration.parse(
+        DeclarativeConfiguration.parse(
             new ByteArrayInputStream(yaml.getBytes(StandardCharsets.UTF_8)), envVars);
     assertThat(model)
         .isEqualTo(