+++ sdk-extensions/incubator/src/test/java/io/opentelemetry/sdk/extension/incubator/fileconfig/component/ResourceOrderedFirstComponentProvider.java
@@ -5,9 +5,9 @@
 
 package io.opentelemetry.sdk.extension.incubator.fileconfig.component;
 
+import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.Ordered;
 import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
-import io.opentelemetry.sdk.autoconfigure.spi.internal.StructuredConfigProperties;
 import io.opentelemetry.sdk.resources.Resource;
 
 public class ResourceOrderedFirstComponentProvider implements ComponentProvider<Resource>, Ordered {
@@ -22,7 +22,7 @@ public class ResourceOrderedFirstComponentProvider implements ComponentProvider<
   }
 
   @Override
-  public Resource create(StructuredConfigProperties config) {
+  public Resource create(DeclarativeConfigProperties config) {
     return Resource.builder().put("order", "first").build();
   }