+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/export/BatchLogRecordProcessorBuilder.java
@@ -11,6 +11,8 @@ import static java.util.Objects.requireNonNull;
 import io.opentelemetry.api.metrics.MeterProvider;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Builder class for {@link BatchLogRecordProcessor}.
@@ -18,6 +20,8 @@ import java.util.concurrent.TimeUnit;
  * @since 1.27.0
  */
 public final class BatchLogRecordProcessorBuilder {
+  private static final Logger logger =
+      Logger.getLogger(BatchLogRecordProcessorBuilder.class.getName());
 
   // Visible for testing
   static final long DEFAULT_SCHEDULE_DELAY_MILLIS = 1000;
@@ -150,6 +154,13 @@ public final class BatchLogRecordProcessorBuilder {
    * @return a new {@link BatchLogRecordProcessor}.
    */
   public BatchLogRecordProcessor build() {
+    if (maxExportBatchSize > maxQueueSize) {
+      logger.log(
+          Level.WARNING,
+          "maxExportBatchSize should not exceed maxQueueSize. Setting maxExportBatchSize to {0} instead of {1}",
+          new Object[] {maxQueueSize, maxExportBatchSize});
+      maxExportBatchSize = maxQueueSize;
+    }
     return new BatchLogRecordProcessor(
         logRecordExporter,
         meterProvider,