+++ sdk/logs/src/test/java/io/opentelemetry/sdk/logs/SdkLogRecordBuilderTest.java
@@ -5,7 +5,12 @@
 
 package io.opentelemetry.sdk.logs;
 
+import static io.opentelemetry.api.common.AttributeKey.booleanKey;
+import static io.opentelemetry.api.common.AttributeKey.doubleKey;
+import static io.opentelemetry.api.common.AttributeKey.longKey;
+import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
+import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
 import static org.mockito.Mockito.when;
 
 import io.opentelemetry.api.common.AttributeKey;
@@ -75,7 +80,7 @@ class SdkLogRecordBuilderTest {
     builder.setTimestamp(timestamp);
     builder.setObservedTimestamp(456, TimeUnit.SECONDS);
     builder.setObservedTimestamp(observedTimestamp);
-    builder.setAttribute(null, null);
+    builder.setAttribute((String) null, (String) null);
     builder.setAttribute(AttributeKey.stringKey("k1"), "v1");
     builder.setAllAttributes(Attributes.builder().put("k2", "v2").put("k3", "v3").build());
     builder.setContext(Span.wrap(spanContext).storeInContext(Context.root()));
@@ -113,5 +118,23 @@ class SdkLogRecordBuilderTest {
         .hasAttributes(Attributes.empty())
         .hasSpanContext(SpanContext.getInvalid())
         .hasSeverity(Severity.UNDEFINED_SEVERITY_NUMBER);
+  }
+
+  @Test
+  void testConvenienceAttributeMethods() {
+    builder
+        .setAttribute("foo", "bar")
+        .setAttribute("lk", 12L)
+        .setAttribute("dk", 12.123)
+        .setAttribute("bk", true)
+        .setAttribute("ik", 13)
+        .emit();
+    assertThat(emittedLog.get().toLogRecordData())
+        .hasAttributesSatisfyingExactly(
+            equalTo(stringKey("foo"), "bar"),
+            equalTo(longKey("lk"), 12L),
+            equalTo(doubleKey("dk"), 12.123),
+            equalTo(booleanKey("bk"), true),
+            equalTo(longKey("ik"), 13L));
   }
 }